ai-service/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Deep Learning - PyTorch
torch==2.5.1
torchvision==0.20.1

# OCR
easyocr==1.7.2

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6

# CORS
aiofiles==24.1.0
